+++ src/ingest_validation_tools/schema_loader.py
@@ -8,24 +8,29 @@ _table_schemas_path = Path(__file__).parent / 'table-schemas'
 _directory_schemas_path = Path(__file__).parent / 'directory-schemas'
 
 
+class PreflightError(Exception):
+    pass
+
+
 SchemaVersion = namedtuple('SchemaVersion', ['schema_name', 'version'])
 
 
 def get_schema_version_from_row(path, row):
     '''
-    >>> _get_schema_version_from_row('empty', {'bad-column': 'bad-value'})
+    >>> get_schema_version_from_row('empty', {'bad-column': 'bad-value'})
     Traceback (most recent call last):
     ...
-    submission.PreflightError: empty does not contain "assay_type". Column headers: bad-column
+    schema_loader.PreflightError: empty does not contain "assay_type". Column headers: bad-column
 
-    >>> _get_schema_version_from_row('v0', {'assay_type': 'PAS microscopy'})
+    >>> get_schema_version_from_row('v0', {'assay_type': 'PAS microscopy'})
     SchemaVersion(schema_name='stained', version=0)
 
-    >>> _get_schema_version_from_row('v42', {'assay_type': 'PAS microscopy', 'version': 42})
+    >>> get_schema_version_from_row('v42', {'assay_type': 'PAS microscopy', 'version': 42})
     SchemaVersion(schema_name='stained', version=42)
 
-    >>> _get_schema_version_from_row('xyz-v42', {'assay_type': 'PAS microscopy', 'version': 42, 'source_project': 'XYZ'})
-    SchemaVersion(schema_name='stained-XYZ', version=42)
+    >>> get_schema_version_from_row('xyz-v42',
+    ...     {'assay_type': 'PAS microscopy', 'version': 42, 'source_project': 'XYZ'})
+    SchemaVersion(schema_name='stained (XYZ)', version=42)
 
     '''
     if 'assay_type' not in row:
@@ -42,9 +47,26 @@ def get_schema_version_from_row(path, row):
     version = row['version'] if 'version' in row else 0
     schema_name = _assay_to_schema_name(assay)
     if 'source_project' in row:
-        schema_name += f"-{row['source_project']}"
+        schema_name += f" ({row['source_project']})"
 
     return SchemaVersion(schema_name, version)
+
+
+def _assay_to_schema_name(name):
+    '''
+    Given an assay name, read all the schemas until one matches.
+    Return the schema name, but not the version.
+
+    >>> _assay_to_schema_name('PAS microscopy')
+    'stained'
+
+    '''
+    for path in (Path(__file__).parent / 'table-schemas' / 'assays').glob('*.yaml'):
+        schema = load_yaml(path)
+        for field in schema['fields']:
+            if field['name'] == 'assay_type' and name in field['constraints']['enum']:
+                return path.stem.split('-v')[0]
+    raise PreflightError(f"Can't find schema where '{name}' is in the enum for assay_type")
 
 
 def list_schema_versions():
